% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixmass.R
\name{plot.nixmass}
\alias{plot.nixmass}
\title{Plot modeled SWE values of a nixmass object.}
\usage{
\method{plot}{nixmass}(x, title = NULL, ...)
}
\arguments{
\item{x}{nixmass object.}

\item{title}{Main plot title.}

\item{...}{Further graphical parameters may also be supplied as arguments. See \code{\link[graphics]{plot}}.}
}
\value{
Does not return anything. A plot is produced.
}
\description{
Plot modeled SWE values of a nixmass object.
}
\examples{
data("hsdata")
plot(nixmass(hsdata, model = "delta.snow"))

}
