\name{plot.nixmass}
\alias{plot.nixmass}
\title{Plot modeled SWE values of a nixmass object.}
\usage{
\method{plot}{nixmass}(x, title = NULL, ...)
}
\arguments{
     \item{x}{nixmass object.}
     \item{title}{Main plot title.}
     \item{...}{Further graphical parameters may also be supplied as arguments. See \code{\link[graphics]{plot}}.}
}
\description{
     Plot modeled SWE values of a nixmass object. 
}
