% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixmass.R
\name{summary.nixmass}
\alias{summary.nixmass}
\title{Print summary of a nixmass object.}
\usage{
\method{summary}{nixmass}(object, ...)
}
\arguments{
\item{object}{A nixmass object.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Summary information of SWE values calculated with selected models is printed to the screen.
}
\description{
Print summary of a nixmass object.
}
\examples{
data("hsdata")
n <- nixmass(hsdata, model = c("delta.snow", "pi16"))
summary(n)

}
