\name{summary.nixmass}
\alias{summary.nixmass}
\title{Print summary of a nixmass object.}
\usage{
\method{summary}{nixmass}(object, ...)
}
\arguments{
     \item{object}{nixmass object.}
     \item{...}{additional arguments affecting the summary produced.}
}
\description{
     Print summary of a nixmass object. 
}
